+++ app/routes/components/CodeBlock/index.tsx
@@ -1,11 +1,15 @@
-import { ComponentTemplate } from "src/components/ComponentTemplate";
-import { Section } from "src/components/Section";
+import { useState } from "react";
+import { Typography } from "@/components/Typography";
+import { DocLayout } from "@/layouts/DocLayout/doc-layout";
 import type { Route } from "./+types/index";
 import { changelog } from "./changelog.codegen";
-import { Demo } from "./Demo";
-import { demoSourceCode } from "./Demo_source_code.codegen";
+import Demo01 from "./demos/Demo01";
+import demo01SourceCode from "./demos/Demo01.source_code.codegen";
+import Demo02 from "./demos/Demo02";
+import demo02SourceCode from "./demos/Demo02.source_code.codegen";
+import Demo03 from "./demos/Demo03";
+import demo03SourceCode from "./demos/Demo03.source_code.codegen";
 import { sourceCode } from "./source_code.codegen";
-import { Typography } from "@/components/Typography";
 
 export function meta(_: Route.MetaArgs) {
   return [
@@ -18,20 +22,43 @@ export function meta(_: Route.MetaArgs) {
 }
 
 export default function CodeBlockPage() {
+  const [tabsValue, setTabsValue] = useState("doc");
+
+  const tabs = [
+    { name: "doc", label: "Document" },
+    { name: "source", label: "Source Code" },
+    { name: "changelog", label: "Changelog" },
+  ];
+
   return (
-    <ComponentTemplate component="CodeBlock">
-      <Section.Root title="CodeBlock">
-        <Section.Demo node={<Demo />} code={demoSourceCode} />
-        <Typography.H2 mt="1.5rem">Usage</Typography.H2>
-        <Typography.P>
-          The CodeBlock component is just built on top of the ScrollArea and
-          CodeHighlight - very simple. It give you a better component to show
-          your code snippets with a max height and a scroll when needed - and
-          good background color (gray-000).
-        </Typography.P>
-        <Section.SourceCode files={sourceCode} />
-        <Section.Changelog changelog={changelog} />
-      </Section.Root>
-    </ComponentTemplate>
+    <DocLayout>
+      <DocLayout.Title
+        title="CodeBlock"
+        desc="For displaying code snippets."
+        tabsValue={tabsValue}
+        tabs={tabs}
+        onTabsValueChange={setTabsValue}
+      />
+      {tabsValue === "doc" && (
+        <DocLayout.Content>
+          <DocLayout.Live node={<Demo01 />} code={demo01SourceCode} />
+          <Typography.H2>Examples</Typography.H2>
+          <Typography.H3>With Title</Typography.H3>
+          <DocLayout.Live node={<Demo02 />} code={demo02SourceCode} />
+          <Typography.H3>Without height or maxHeight</Typography.H3>
+          <DocLayout.Live node={<Demo03 />} code={demo03SourceCode} />
+        </DocLayout.Content>
+      )}
+      {tabsValue === "source" && (
+        <DocLayout.Content>
+          <DocLayout.Files files={sourceCode} />
+        </DocLayout.Content>
+      )}
+      {tabsValue === "changelog" && (
+        <DocLayout.Content>
+          <DocLayout.Changelog changelog={changelog} />
+        </DocLayout.Content>
+      )}
+    </DocLayout>
   );
 }