+++ src/components/CodeBlock/index.ts
@@ -1,3 +1,3 @@
 // CodeBlock from copy-ui @ 2025-10-12.
 
-export { CodeBlock } from "./CodeBlock";
+export { CodeBlock } from "./code-block";