+++ src/components/CodeBlock/CHANGELOG.md
@@ -1,2 +1,4 @@
 - 2025-10-09: Initial version.
 - 2025-10-12: Support the `className` prop.
+- 2025-11-03: Support the `title` prop. Rename the `height` prop to
+  `scrollAreaHeight`, and rename the `maxHeight` prop to `scrollAreaMaxHeight`.